+++ pkg/translator/prometheusremotewrite/helper.go
@@ -46,6 +46,10 @@ const (
 	// according to the prometheus specification
 	// https://github.com/OpenObservability/OpenMetrics/blob/main/specification/OpenMetrics.md#exemplars
 	maxExemplarRunes = 128
+	// Trace and Span id keys are defined as part of the spec:
+	// https://github.com/open-telemetry/opentelemetry-specification/blob/main/specification%2Fmetrics%2Fdatamodel.md#exemplars-2
+	traceIDKey = "trace_id"
+	spanIDKey  = "span_id"
 )
 
 type bucketBoundsData struct {
@@ -357,13 +361,32 @@ func getPromExemplars(pt pdata.HistogramDataPoint) []prompb.Exemplar {
 
 	for i := 0; i < pt.Exemplars().Len(); i++ {
 		exemplar := pt.Exemplars().At(i)
+		exemplarRunes := 0
 
 		promExemplar := &prompb.Exemplar{
 			Value:     exemplar.DoubleVal(),
 			Timestamp: timestamp.FromTime(exemplar.Timestamp().AsTime()),
 		}
+		if !exemplar.TraceID().IsEmpty() {
+			val := exemplar.TraceID().HexString()
+			exemplarRunes += utf8.RuneCountInString(traceIDKey) + utf8.RuneCountInString(val)
+			promLabel := prompb.Label{
+				Name:  traceIDKey,
+				Value: val,
+			}
+			promExemplar.Labels = append(promExemplar.Labels, promLabel)
+		}
+		if !exemplar.SpanID().IsEmpty() {
+			val := exemplar.SpanID().HexString()
+			exemplarRunes += utf8.RuneCountInString(spanIDKey) + utf8.RuneCountInString(val)
+			promLabel := prompb.Label{
+				Name:  spanIDKey,
+				Value: val,
+			}
+			promExemplar.Labels = append(promExemplar.Labels, promLabel)
+		}
+		var labelsFromAttributes []prompb.Label
 
-		exemplarRunes := 0
 		exemplar.FilteredAttributes().Range(func(key string, value pdata.AttributeValue) bool {
 			val := value.AsString()
 			exemplarRunes += utf8.RuneCountInString(key) + utf8.RuneCountInString(val)
@@ -372,12 +395,14 @@ func getPromExemplars(pt pdata.HistogramDataPoint) []prompb.Exemplar {
 				Value: val,
 			}
 
-			promExemplar.Labels = append(promExemplar.Labels, promLabel)
+			labelsFromAttributes = append(labelsFromAttributes, promLabel)
 
 			return true
 		})
-		if exemplarRunes > maxExemplarRunes {
-			promExemplar.Labels = nil
+		if exemplarRunes <= maxExemplarRunes {
+			// only append filtered attributes if it does not cause exemplar
+			// labels to exceed the max number of runes
+			promExemplar.Labels = append(promExemplar.Labels, labelsFromAttributes...)
 		}
 
 		promExemplars = append(promExemplars, *promExemplar)